+++ sdk/resourcemanager/redhatopenshift/armredhatopenshift/fake/server_factory.go
@@ -19,13 +19,14 @@ import (
 
 // ServerFactory is a fake server for instances of the armredhatopenshift.ClientFactory type.
 type ServerFactory struct {
-	MachinePoolsServer          MachinePoolsServer
-	OpenShiftClustersServer     OpenShiftClustersServer
-	OpenShiftVersionsServer     OpenShiftVersionsServer
-	OperationsServer            OperationsServer
-	SecretsServer               SecretsServer
-	SyncIdentityProvidersServer SyncIdentityProvidersServer
-	SyncSetsServer              SyncSetsServer
+	MachinePoolsServer                     MachinePoolsServer
+	OpenShiftClustersServer                OpenShiftClustersServer
+	OpenShiftVersionsServer                OpenShiftVersionsServer
+	OperationsServer                       OperationsServer
+	PlatformWorkloadIdentityRoleSetsServer PlatformWorkloadIdentityRoleSetsServer
+	SecretsServer                          SecretsServer
+	SyncIdentityProvidersServer            SyncIdentityProvidersServer
+	SyncSetsServer                         SyncSetsServer
 }
 
 // NewServerFactoryTransport creates a new instance of ServerFactoryTransport with the provided implementation.
@@ -40,15 +41,16 @@ func NewServerFactoryTransport(srv *ServerFactory) *ServerFactoryTransport {
 // ServerFactoryTransport connects instances of armredhatopenshift.ClientFactory to instances of ServerFactory.
 // Don't use this type directly, use NewServerFactoryTransport instead.
 type ServerFactoryTransport struct {
-	srv                           *ServerFactory
-	trMu                          sync.Mutex
-	trMachinePoolsServer          *MachinePoolsServerTransport
-	trOpenShiftClustersServer     *OpenShiftClustersServerTransport
-	trOpenShiftVersionsServer     *OpenShiftVersionsServerTransport
-	trOperationsServer            *OperationsServerTransport
-	trSecretsServer               *SecretsServerTransport
-	trSyncIdentityProvidersServer *SyncIdentityProvidersServerTransport
-	trSyncSetsServer              *SyncSetsServerTransport
+	srv                                      *ServerFactory
+	trMu                                     sync.Mutex
+	trMachinePoolsServer                     *MachinePoolsServerTransport
+	trOpenShiftClustersServer                *OpenShiftClustersServerTransport
+	trOpenShiftVersionsServer                *OpenShiftVersionsServerTransport
+	trOperationsServer                       *OperationsServerTransport
+	trPlatformWorkloadIdentityRoleSetsServer *PlatformWorkloadIdentityRoleSetsServerTransport
+	trSecretsServer                          *SecretsServerTransport
+	trSyncIdentityProvidersServer            *SyncIdentityProvidersServerTransport
+	trSyncSetsServer                         *SyncSetsServerTransport
 }
 
 // Do implements the policy.Transporter interface for ServerFactoryTransport.
@@ -80,6 +82,11 @@ func (s *ServerFactoryTransport) Do(req *http.Request) (*http.Response, error) {
 	case "OperationsClient":
 		initServer(s, &s.trOperationsServer, func() *OperationsServerTransport { return NewOperationsServerTransport(&s.srv.OperationsServer) })
 		resp, err = s.trOperationsServer.Do(req)
+	case "PlatformWorkloadIdentityRoleSetsClient":
+		initServer(s, &s.trPlatformWorkloadIdentityRoleSetsServer, func() *PlatformWorkloadIdentityRoleSetsServerTransport {
+			return NewPlatformWorkloadIdentityRoleSetsServerTransport(&s.srv.PlatformWorkloadIdentityRoleSetsServer)
+		})
+		resp, err = s.trPlatformWorkloadIdentityRoleSetsServer.Do(req)
 	case "SecretsClient":
 		initServer(s, &s.trSecretsServer, func() *SecretsServerTransport { return NewSecretsServerTransport(&s.srv.SecretsServer) })
 		resp, err = s.trSecretsServer.Do(req)